+++ synphot/models.py
@@ -2,6 +2,7 @@
 """Spectrum models not in `astropy.modeling`."""
 
 # STDLIB
+import math
 import warnings
 from copy import deepcopy
 from functools import partial
@@ -134,6 +135,12 @@ class BlackBody1D(Fittable1DModel):
 
         return bbflux.value
 
+    def integrate(self, *args):
+        with u.add_enabled_equivalencies(u.temperature()):
+            t = u.Quantity(self.temperature, u.K)
+
+        return (const.sigma_sb * t ** 4 / math.pi)  # per steradian
+
 
 class BlackBodyNorm1D(BlackBody1D):
     """Create a normalized :ref:`blackbody spectrum <synphot-planck-law>`
@@ -176,6 +183,9 @@ class BlackBodyNorm1D(BlackBody1D):
         bbflux = super(BlackBodyNorm1D, self).evaluate(x, temperature)
         return bbflux * self._omega
 
+    def integrate(self, *args):
+        return super().integrate(*args) * self._omega
+
 
 class Box1D(_models.Box1D):
     """Same as `astropy.modeling.functional_models.Box1D`, except with
@@ -214,6 +224,13 @@ class Box1D(_models.Box1D):
                 self._calc_sampleset, step=step, minimal=minimal), w1, w2))
 
         return np.asarray(w)
+
+    def integrate(self, *args):
+        # TODO: Remove unit hardcoding when we use model with units natively.
+        with u.add_enabled_equivalencies(u.spectral()):
+            w = u.Quantity(self.width, u.AA)
+
+        return self.amplitude * w
 
 
 class ConstFlux1D(_models.Const1D):
@@ -268,6 +285,19 @@ class ConstFlux1D(_models.Const1D):
         a = (self.amplitude * np.ones_like(x)) * self._flux_unit
         y = units.convert_flux(x, a, units.PHOTLAM)
         return y.value
+
+    def integrate(self, x):
+        # TODO: Remove unit hardcoding when we use model with units natively.
+        # TODO: We do not handle wav_unit as wave number nor energy for now.
+        if 'wav' in self._flux_unit.physical_type:
+            wav_unit = u.AA
+        else:
+            wav_unit = u.Hz
+        with u.add_enabled_equivalencies(u.spectral()):
+            x = u.Quantity(x, wav_unit)
+            amp = u.Quantity(self.amplitude, self._flux_unit)
+
+        return (max(x) - min(x)) * amp
 
 
 class Empirical1D(Tabular1D):
@@ -414,6 +444,8 @@ class BaseGaussian1D(_models.Gaussian1D):
     ``sampleset`` defined.
 
     """
+    _sqrt_2_pi = math.sqrt(2 * math.pi)
+
     def sampleset(self, factor_step=0.1, **kwargs):
         """Return ``x`` array that samples the feature.
 
@@ -448,9 +480,16 @@ class Gaussian1D(BaseGaussian1D):
     ``sampleset`` defined.
 
     """
-    pass
+    def integrate(self, *args):
+        # TODO: Remove unit hardcoding when we use model with units natively.
+        with u.add_enabled_equivalencies(u.spectral()):
+            stddev = u.Quantity(self.stddev, u.AA)
+
+        return self.amplitude * stddev * self._sqrt_2_pi
 
 
+# TODO: Deprecate this?
+# This is not really supported anymore but kept for backward compatibility.
 class GaussianAbsorption1D(BaseGaussian1D):
     """Same as ``astropy.modeling.functional_models.GaussianAbsorption1D``,
     except with ``sampleset`` defined.
@@ -509,7 +548,7 @@ class GaussianFlux1D(Gaussian1D):
         else:
             self.stddev = fwhm * gaussian_fwhm_to_sigma
 
-        gaussian_amp_to_totflux = np.sqrt(2.0 * np.pi) * self.stddev
+        gaussian_amp_to_totflux = self._sqrt_2_pi * self.stddev
 
         if total_flux is None:
             u_str = 'PHOTLAM'
@@ -528,6 +567,10 @@ class GaussianFlux1D(Gaussian1D):
 
         self.meta['expr'] = 'em({0:g}, {1:g}, {2:g}, {3})'.format(
             self.mean.value, fwhm, total_flux, u_str)
+
+    def integrate(self, *args):
+        # TODO: Remove unit hardcoding when we use model with units natively.
+        return super(GaussianFlux1D, self).integrate(*args) * units.PHOTLAM
 
 
 class Lorentz1D(_models.Lorentz1D):
@@ -558,6 +601,18 @@ class Lorentz1D(_models.Lorentz1D):
 
         return np.asarray(w)
 
+    def integrate(self, x):
+        # TODO: Remove unit hardcoding when we use model with units natively.
+        with u.add_enabled_equivalencies(u.spectral()):
+            x = u.Quantity(x, u.AA)
+            x_0 = u.Quantity(self.x_0, u.AA)
+            gamma = u.Quantity(self.fwhm, u.AA) * 0.5
+
+        a1 = np.arctan((min(x) - x_0) / gamma)
+        a2 = np.arctan((max(x) - x_0) / gamma)
+        da = (a2 - a1).to(u.dimensionless_unscaled, u.dimensionless_angles())
+        return self.amplitude * gamma * da
+
 
 class RickerWavelet1D(_RickerWavelet1D):
     """Same as `astropy.modeling.functional_models.RickerWavelet1D`, except
@@ -586,6 +641,37 @@ class RickerWavelet1D(_RickerWavelet1D):
             w = list(map(np.arange, w1, w2, dw))
 
         return np.asarray(w)
+
+    def integrate(self, x):
+        # TODO: Remove unit hardcoding when we use model with units natively.
+        with u.add_enabled_equivalencies(u.spectral()):
+            x = u.Quantity(x, u.AA)
+            x_0 = u.Quantity(self.x_0, u.AA)
+            sig = u.Quantity(self.sigma, u.AA)
+
+        # Roots, where y=0
+        root_left = x_0 - sig
+        root_right = x_0 + sig
+
+        x_min = min(x)
+        x_max = max(x)
+        if x_min >= root_left or x_max <= root_right:
+            raise NotImplementedError(
+                'Partial analytic integration not supported')
+
+        sig2 = sig * sig
+
+        def _int_subregion(xx1, xx2):
+            dx_min = xx1 - x_0
+            dx_max = xx2 - x_0
+            a1 = dx_min * np.exp(-0.5 * dx_min * dx_min / sig2)
+            a2 = dx_max * np.exp(-0.5 * dx_max * dx_max / sig2)
+            return abs(a2 - a1)
+
+        # Unsigned area
+        return self.amplitude * (_int_subregion(x_min, root_left) +
+                                 _int_subregion(root_left, root_right) +
+                                 _int_subregion(root_right, x_max))
 
 
 # TODO: Emit proper deprecation warning.
@@ -642,6 +728,17 @@ class PowerLawFlux1D(_models.PowerLaw1D):
         flux = units.convert_flux(x, y, units.PHOTLAM)
         return flux.value
 
+    def integrate(self, x):
+        # TODO: Remove unit hardcoding when we use model with units natively.
+        with u.add_enabled_equivalencies(u.spectral()):
+            x = u.Quantity(x, u.AA)
+            x_0 = u.Quantity(self.x_0, u.AA)
+            amp = u.Quantity(self.amplitude, self._flux_unit)
+
+        fac = 1 - self.alpha
+        denom = x_0 ** -self.alpha * fac
+        return amp * (max(x) ** fac - min(x) ** fac) / denom
+
 
 class Trapezoid1D(_models.Trapezoid1D):
     """Same as `astropy.modeling.functional_models.Trapezoid1D`, except with
@@ -661,6 +758,14 @@ class Trapezoid1D(_models.Trapezoid1D):
             w = list(zip(x1, x2, x3, x4))
 
         return np.asarray(w)
+
+    def integrate(self, *args):
+        # TODO: Remove unit hardcoding when we use model with units natively.
+        with u.add_enabled_equivalencies(u.spectral()):
+            width = u.Quantity(self.width, u.AA)
+            slope = u.Quantity(self.slope, 1 / u.AA)
+
+        return self.amplitude * (width + self.amplitude / slope)
 
 
 # Functions below are for sampleset magic.