+++ synphot/tests/test_units.py
@@ -3,6 +3,8 @@
 
 .. note:: VEGAMAG conversion is tested in test_spectrum.py.
 
+.. note:: spectral_density_integrated is tested in astropy>=4.1.
+
 """
 
 # THIRD-PARTY